git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@2414 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/paging.cpp
@@ -248,7 +248,7 @@ PIGGY_PAGE_IN (gameData.pig.tex.bmIndexP [m_nBaseTex].index, gameStates.app.bD1D
 
 //------------------------------------------------------------------------------
 
-void CSegment::LoadSideTextures (short nSide)
+void CSegment::LoadSideTextures (int nSide)
 {
 #if DBG
 if ((Index () == nDbgSeg) && ((nDbgSide < 0) || (nSide == nDbgSide)))
